+++ CHANGES.rst
@@ -8,6 +8,9 @@ Bug fix:
  * MXSecretShareManager: Fix crash in cancelRequestWithRequestId (vector-im/riot-ios/issues/3272).
  * MXIdentityService: Fix crash in handleHTTPClientError (vector-im/riot-ios/issues/3273).
 
+API break:
+ * MXCrossSigning: Removed MXCrossSigningStateCanCrossSignAsynchronously.
+
 Changes in Matrix iOS SDK in 0.16.5 (2020-05-18)
 ================================================